Embed EblStateMachine into ShippingInstruction
