+++ edocumentation-domain/src/test/java/org/dcsa/edocumentation/domain/ebl/EblStateMachineTest.java
@@ -1,5 +1,6 @@
 package org.dcsa.edocumentation.domain.ebl;
 
+import org.dcsa.edocumentation.domain.persistence.entity.ShippingInstruction;
 import org.dcsa.edocumentation.domain.persistence.entity.enums.EblDocumentStatus;
 import org.dcsa.skernel.errors.exceptions.ConflictException;
 import org.dcsa.skernel.errors.exceptions.InternalServerErrorException;
@@ -7,6 +8,7 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
+import java.util.UUID;
 import java.util.function.Function;
 
 import static org.dcsa.edocumentation.domain.persistence.entity.enums.EblDocumentStatus.*;
@@ -16,82 +18,88 @@ class EblStateMachineTest {
 
   @Test
   void testBasicDefaultFlow() {
-    EBLStateMachine stateMachine = EBLStateMachine.fromInitialStateUsingDefaultFlow();
-    Assertions.assertEquals(RECE, stateMachine.getCurrentStatus());
+    ShippingInstruction stateMachine = ShippingInstruction.builder().build();
+    Assertions.assertEquals(RECE, stateMachine.getDocumentStatus());
     Assertions.assertTrue(stateMachine.isPendingUpdateSupported());
     stateMachine.pendingUpdate();
-    Assertions.assertEquals(PENU, stateMachine.getCurrentStatus());
+    Assertions.assertEquals(PENU, stateMachine.getDocumentStatus());
     stateMachine.receive();
-    Assertions.assertEquals(RECE, stateMachine.getCurrentStatus());
+    Assertions.assertEquals(RECE, stateMachine.getDocumentStatus());
     stateMachine.draft();
-    Assertions.assertEquals(DRFT, stateMachine.getCurrentStatus());
+    Assertions.assertEquals(DRFT, stateMachine.getDocumentStatus());
 
     // Verify resume
-    stateMachine = EBLStateMachine.resumeFromStateUsingDefaultFlow(DRFT);
-    Assertions.assertEquals(DRFT, stateMachine.getCurrentStatus());
+    stateMachine = ShippingInstruction.builder().documentStatus(DRFT).build();
+    Assertions.assertEquals(DRFT, stateMachine.getDocumentStatus());
     Assertions.assertTrue(stateMachine.isPendingApprovalSupported());
     stateMachine.pendingApproval();
-    Assertions.assertEquals(PENA, stateMachine.getCurrentStatus());
+    Assertions.assertEquals(PENA, stateMachine.getDocumentStatus());
     stateMachine.approve();
-    Assertions.assertEquals(APPR, stateMachine.getCurrentStatus());
+    Assertions.assertEquals(APPR, stateMachine.getDocumentStatus());
     stateMachine.issue();
-    Assertions.assertEquals(ISSU, stateMachine.getCurrentStatus());
+    Assertions.assertEquals(ISSU, stateMachine.getDocumentStatus());
   }
 
   @Test
   void testBasicAmendmentFlow() {
-    EBLStateMachine stateMachine = EBLStateMachine.fromInitialStateUsingAmendmentFlow();
-    Assertions.assertEquals(RECE, stateMachine.getCurrentStatus());
+    ShippingInstruction stateMachine = ShippingInstruction.builder().amendmentToTransportDocument(UUID.randomUUID())
+      .build();
+    Assertions.assertEquals(RECE, stateMachine.getDocumentStatus());
     stateMachine.draft();
-    Assertions.assertEquals(DRFT, stateMachine.getCurrentStatus());
+    Assertions.assertEquals(DRFT, stateMachine.getDocumentStatus());
     stateMachine.approve();
-    Assertions.assertEquals(APPR, stateMachine.getCurrentStatus());
+    Assertions.assertEquals(APPR, stateMachine.getDocumentStatus());
     stateMachine.issue();
-    Assertions.assertEquals(ISSU, stateMachine.getCurrentStatus());
+    Assertions.assertEquals(ISSU, stateMachine.getDocumentStatus());
 
-    stateMachine = EBLStateMachine.resumeFromStateUsingAmendmentFlow(DRFT);
-    Assertions.assertEquals(DRFT, stateMachine.getCurrentStatus());
+    stateMachine = ShippingInstruction.builder().amendmentToTransportDocument(UUID.randomUUID())
+      .documentStatus(DRFT)
+      .build();
+    Assertions.assertEquals(DRFT, stateMachine.getDocumentStatus());
     stateMachine.approve();
-    Assertions.assertEquals(APPR, stateMachine.getCurrentStatus());
+    Assertions.assertEquals(APPR, stateMachine.getDocumentStatus());
   }
 
 
   @Test
   void testBasicAmendmentFlowRestricted() {
-    EBLStateMachine stateMachine = EBLStateMachine.fromInitialStateUsingAmendmentFlow();
-    Assertions.assertEquals(RECE, stateMachine.getCurrentStatus());
+    ShippingInstruction stateMachine = ShippingInstruction.builder().amendmentToTransportDocument(UUID.randomUUID())
+      .build();
+    Assertions.assertEquals(RECE, stateMachine.getDocumentStatus());
 
     // RECE must go directly to DRFT in the amendment flow (which is weird because there is no
     // reject state either... >.>)
     Assertions.assertFalse(stateMachine.isPendingUpdateSupported());
     Assertions.assertThrows(InternalServerErrorException.class, stateMachine::pendingUpdate);
     // The error should not have made it change state
-    Assertions.assertEquals(RECE, stateMachine.getCurrentStatus());
+    Assertions.assertEquals(RECE, stateMachine.getDocumentStatus());
 
     stateMachine.draft();
     // DRFT must go directly to APPR in the amendment flow.
     Assertions.assertFalse(stateMachine.isPendingApprovalSupported());
     Assertions.assertThrows(InternalServerErrorException.class, stateMachine::pendingApproval);
     // Again, it retains the current state on invalid transitions.
-    Assertions.assertEquals(DRFT, stateMachine.getCurrentStatus());
+    Assertions.assertEquals(DRFT, stateMachine.getDocumentStatus());
 
 
-    stateMachine = EBLStateMachine.resumeFromStateUsingAmendmentFlow(DRFT);
+    stateMachine = ShippingInstruction.builder().amendmentToTransportDocument(UUID.randomUUID())
+      .documentStatus(DRFT)
+      .build();
     // DRFT must go directly to APPR in the amendment flow - even after resume
     Assertions.assertFalse(stateMachine.isPendingApprovalSupported());
     Assertions.assertThrows(InternalServerErrorException.class, stateMachine::pendingApproval);
     // Again, it retains the current state on invalid transitions.
-    Assertions.assertEquals(DRFT, stateMachine.getCurrentStatus());
+    Assertions.assertEquals(DRFT, stateMachine.getDocumentStatus());
   }
 
   @Test
   void loadInvalidStateTransitions() {
-    List<Function<EblDocumentStatus, EBLStateMachine>> initializers = List.of(
-      EBLStateMachine::resumeFromStateUsingDefaultFlow,
-      EBLStateMachine::resumeFromStateUsingAmendmentFlow
+    List<Function<EblDocumentStatus, ShippingInstruction>> initializers = List.of(
+      (s) -> ShippingInstruction.builder().documentStatus(s).build(),
+      (s) -> ShippingInstruction.builder().amendmentToTransportDocument(UUID.randomUUID()).documentStatus(s).build()
     );
-    for (Function<EblDocumentStatus, EBLStateMachine> initializer : initializers) {
-      EBLStateMachine stateMachine = initializer.apply(VOID);
+    for (Function<EblDocumentStatus, ShippingInstruction> initializer : initializers) {
+      ShippingInstruction stateMachine = initializer.apply(VOID);
       Assertions.assertThrows(ConflictException.class, stateMachine::receive);
       Assertions.assertThrows(ConflictException.class, stateMachine::pendingUpdate);
       Assertions.assertThrows(ConflictException.class, stateMachine::draft);
